Группирует служебные функции
