+++ src/utils/keyboard.js
@@ -1,0 +1,3 @@
+import {ESC_KEY} from "../const";
+
+export const checkEscKey = (key) => key === ESC_KEY;
